Struct x509_parser::certificate::TbsCertificate

pub struct TbsCertificate<'a> {
    pub version: X509Version,
    pub serial: BigUint,
    pub signature: AlgorithmIdentifier<'a>,
    pub issuer: X509Name<'a>,
    pub validity: Validity,
    pub subject: X509Name<'a>,
    pub subject_pki: SubjectPublicKeyInfo<'a>,
    pub issuer_uid: Option<UniqueIdentifier<'a>>,
    pub subject_uid: Option<UniqueIdentifier<'a>>,
    // some fields omitted
}

The sequence TBSCertificate contains information associated with the subject of the certificate and the CA that issued it.

RFC5280 definition:

  TBSCertificate  ::=  SEQUENCE  {
       version         [0]  EXPLICIT Version DEFAULT v1,
       serialNumber         CertificateSerialNumber,
       signature            AlgorithmIdentifier,
       issuer               Name,
       validity             Validity,
       subject              Name,
       subjectPublicKeyInfo SubjectPublicKeyInfo,
       issuerUniqueID  [1]  IMPLICIT UniqueIdentifier OPTIONAL,
                            -- If present, version MUST be v2 or v3
       subjectUniqueID [2]  IMPLICIT UniqueIdentifier OPTIONAL,
                            -- If present, version MUST be v2 or v3
       extensions      [3]  EXPLICIT Extensions OPTIONAL
                            -- If present, version MUST be v3
       }

Fields
version: X509Versionserial: BigUintsignature: AlgorithmIdentifier<'a>issuer: X509Name<'a>validity: Validitysubject: X509Name<'a>subject_pki: SubjectPublicKeyInfo<'a>issuer_uid: Option<UniqueIdentifier<'a>>subject_uid: Option<UniqueIdentifier<'a>>
Implementations

impl<'a> TbsCertificate<'a>

pub fn extensions(&self) -> &[X509Extension<'a>]

Returns the certificate extensions

pub fn iter_extensions(&self) -> impl Iterator<Item = &X509Extension<'a>>

Returns an iterator over the certificate extensions

pub fn find_extension(&self, oid: &Oid<'_>) -> Option<&X509Extension<'a>>

Searches for an extension with the given Oid.

Note: if there are several extensions with the same Oid, the first one is returned.

pub fn extensions_map(
&self
) -> Result<HashMap<Oid<'_>, &X509Extension<'a>>, X509Error>

Builds and returns a map of extensions.

If an extension is present twice, this will fail and return DuplicateExtensions.

pub fn basic_constraints(&self) -> Option<(bool, &BasicConstraints)>

pub fn key_usage(&self) -> Option<(bool, &KeyUsage)>

pub fn extended_key_usage(&self) -> Option<(bool, &ExtendedKeyUsage<'a>)>

pub fn policy_constraints(&self) -> Option<(bool, &PolicyConstraints)>

pub fn inhibit_anypolicy(&self) -> Option<(bool, &InhibitAnyPolicy)>

pub fn policy_mappings(&self) -> Option<(bool, &PolicyMappings<'a>)>

pub fn subject_alternative_name(
&self
) -> Option<(bool, &SubjectAlternativeName<'a>)>

pub fn name_constraints(&self) -> Option<(bool, &NameConstraints<'a>)>

pub fn is_ca(&self) -> bool

Returns true if certificate has basicConstraints CA:true

pub fn raw_serial(&self) -> &'a [u8]

Get the raw bytes of the certificate serial number

pub fn raw_serial_as_string(&self) -> String

Get a formatted string of the certificate serial number, separated by ‘:’