x509_parser::extensions

Enum GeneralName

pub enum GeneralName<'a> {
    OtherName(Oid<'a>, &'a [u8]),
    RFC822Name(&'a str),
    DNSName(&'a str),
    X400Address(Any<'a>),
    DirectoryName(X509Name<'a>),
    EDIPartyName(Any<'a>),
    URI(&'a str),
    IPAddress(&'a [u8]),
    RegisteredID(Oid<'a>),
}

Represents a GeneralName as defined in RFC5280. There is no support X.400 addresses and EDIPartyName.

String formats are not validated.

Variants

OtherName(Oid<'a>, &'a [u8])

RFC822Name(&'a str)

More or less an e-mail, the format is not checked.

DNSName(&'a str)

A hostname, format is not checked.

X400Address(Any<'a>)

X400Address,

DirectoryName(X509Name<'a>)

RFC5280 defines several string types, we always try to parse as utf-8 which is more or less a superset of the string types.

EDIPartyName(Any<'a>)

EDIPartyName

URI(&'a str)

An uniform resource identifier. The format is not checked.

IPAddress(&'a [u8])

An ip address, provided as encoded.

RegisteredID(Oid<'a>)