x509_parser::verify

Function verify_signature

pub fn verify_signature(
    public_key: &SubjectPublicKeyInfo<'_>,
    signature_algorithm: &AlgorithmIdentifier<'_>,
    signature_value: &BitString<'_>,
    raw_data: &[u8],
) -> Result<(), X509Error>

Verify the cryptographic signature of the raw data (can be a certificate, a CRL or a CSR).

public_key is the public key of the signer.

Not all algorithms are supported, this function is limited to what ring supports.