x509_parser::x509

Struct X509Name

pub struct X509Name<'a> { /* private fields */ }

X.509 Name (as used in Issuer and Subject fields)

The Name describes a hierarchical name composed of attributes, such as country name, and corresponding values, such as US. The type of the component AttributeValue is determined by the AttributeType; in general it will be a DirectoryString.

Implementations

impl<'a> X509Name<'a>

pub const fn new( rdn_seq: Vec<RelativeDistinguishedName<'a>>, raw: &'a [u8], ) -> Self

Builds a new X509Name from the provided elements.

pub fn to_string_with_registry( &self, oid_registry: &OidRegistry<'_>, ) -> Result<String, X509Error>

Attempt to format the current name, using the given registry to convert OIDs to strings.

Note: a default registry is provided with this crate, and is returned by the oid_registry() method.

pub fn as_raw(&self) -> &'a [u8]

pub fn iter(&self) -> impl Iterator<Item = &RelativeDistinguishedName<'a>>

Return an iterator over the RelativeDistinguishedName components of the name

pub fn iter_rdn(&self) -> impl Iterator<Item = &RelativeDistinguishedName<'a>>

Return an iterator over the RelativeDistinguishedName components of the name

pub fn iter_attributes( &self, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the attribute types and values of the name

pub fn iter_by_oid( &self, oid: &Oid<'a>, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the components identified by the given OID

The type of the component AttributeValue is determined by the AttributeType; in general it will be a DirectoryString.

Attributes with same OID may be present multiple times, so the returned object is an iterator. Expected number of objects in this iterator are

0: not found

1: present once (common case)

2 or more: attribute is present multiple times

pub fn iter_common_name( &self, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the CommonName attributes of the X.509 Name.

Returned iterator can be empty if there are no CommonName attributes. If you expect only one CommonName to be present, then using next() will get an Option<&AttributeTypeAndValue>.

A common operation is to extract the CommonName as a string.

use x509_parser::x509::X509Name;

fn get_first_cn_as_str<'a>(name: &'a X509Name<'_>) -> Option<&'a str> {
    name.iter_common_name()
        .next()
        .and_then(|cn| cn.as_str().ok())
}

Note that there are multiple reasons for failure or incorrect behavior, for ex. if the attribute is present multiple times, or is not a UTF-8 encoded string (it can be UTF-16, or even an OCTETSTRING according to the standard).

pub fn iter_country(&self) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the Country attributes of the X.509 Name.

pub fn iter_organization( &self, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the Organization attributes of the X.509 Name.

pub fn iter_organizational_unit( &self, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the OrganizationalUnit attributes of the X.509 Name.

pub fn iter_state_or_province( &self, ) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the StateOrProvinceName attributes of the X.509 Name.

pub fn iter_locality(&self) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the Locality attributes of the X.509 Name.

pub fn iter_email(&self) -> impl Iterator<Item = &AttributeTypeAndValue<'a>>

Return an iterator over the EmailAddress attributes of the X.509 Name.