Struct x509_parser::certificate::X509Certificate

pub struct X509Certificate<'a> {
    pub tbs_certificate: TbsCertificate<'a>,
    pub signature_algorithm: AlgorithmIdentifier<'a>,
    pub signature_value: BitStringObject<'a>,
}

An X.509 v3 Certificate.

X.509 v3 certificates are defined in RFC5280, section 4.1. This object uses the same structure for content, so for ex the subject can be accessed using the path x509.tbs_certificate.subject.

X509Certificate also contains convenience methods to access the most common fields (subject, issuer, etc.).

A X509Certificate is a zero-copy view over a buffer, so the lifetime is the same as the buffer containing the binary representation.

fn display_x509_info(x509: &X509Certificate<'_>) {
     let subject = x509.subject();
     let issuer = x509.issuer();
     println!("X.509 Subject: {}", subject);
     println!("X.509 Issuer: {}", issuer);
     println!("X.509 serial: {}", x509.tbs_certificate.raw_serial_as_string());
}

Fields
tbs_certificate: TbsCertificate<'a>signature_algorithm: AlgorithmIdentifier<'a>signature_value: BitStringObject<'a>
Implementations

impl<'a> X509Certificate<'a>

pub fn version(&self) -> X509Version

Get the version of the encoded certificate

pub fn subject(&self) -> &X509Name<'_>

Get the certificate subject.

pub fn issuer(&self) -> &X509Name<'_>

Get the certificate issuer.

pub fn validity(&self) -> &Validity

Get the certificate validity.

pub fn public_key(&self) -> &SubjectPublicKeyInfo<'_>

Get the certificate public key information.

pub fn extensions(&self) -> &[X509Extension<'_>]

Get the certificate extensions.