Module x509_parser::extensions

X.509 Extensions objects and types

Structs

AuthorityInfoAccess

AuthorityKeyIdentifier

BasicConstraints
Identifies whether the subject of the certificate is a CA, and the max validation depth.

CertificatePolicies

ExtendedKeyUsage

GeneralSubtree
Represents the structure used in the name constraints extensions. The fields minimum and maximum are not supported (openssl also has no support).

InhibitAnyPolicy

KeyIdentifier

KeyUsage

NSCertType

NameConstraints

PolicyConstraints

PolicyMappings

SubjectAlternativeName

X509Extension

Enums

GeneralName
Represents a GeneralName as defined in RFC5280. There is no support X.400 addresses and EDIPartyName.

ParsedExtension