Module x509_parser::extensions

X.509 Extensions objects and types

Structs

AccessDescription

AuthorityInfoAccess

AuthorityKeyIdentifier

BasicConstraints

Identifies whether the subject of the certificate is a CA, and the max validation depth.

CRLDistributionPoint

ExtendedKeyUsage

GeneralSubtree

Represents the structure used in the name constraints extensions. The fields minimum and maximum are not supported (openssl also has no support).

InhibitAnyPolicy

KeyIdentifier

KeyUsage

NSCertType

NameConstraints

PolicyConstraints

PolicyInformation

PolicyMapping

PolicyMappings

PolicyQualifierInfo

ReasonFlags

SubjectAlternativeName

UnparsedObject

X509Extension

X.509 version 3 extension

X509ExtensionParser

X509Extension parser builder

Enums

DistributionPointName

GeneralName

Represents a GeneralName as defined in RFC5280. There is no support X.400 addresses and EDIPartyName.

ParsedExtension

Type Definitions

CRLDistributionPoints

CertificatePolicies