x509_parser

Module extensions

X.509 Extensions objects and types

Structs

AccessDescription

AuthorityInfoAccess

AuthorityKeyIdentifier

BasicConstraints

Identifies whether the subject of the certificate is a CA, and the max validation depth.

CRLDistributionPoint

CRLDistributionPoints

CtExtensions

CtExtensions as defined in RFC6962 Section 3.2

CtLogID

LogID as defined in RFC6962 Section 3.2

CtVersion

Certificate Transparency Version as defined in RFC6962 Section 3.2

DigitallySigned

ExtendedKeyUsage

GeneralSubtree

Represents the structure used in the name constraints extensions. The fields minimum and maximum are not supported (openssl also has no support).

InhibitAnyPolicy

IssuerAlternativeName

IssuingDistributionPoint

KeyIdentifier

KeyUsage

NSCertType

NameConstraints

PolicyConstraints

PolicyInformation

PolicyMapping

PolicyMappings

PolicyQualifierInfo

ReasonFlags

SignedCertificateTimestamp

SubjectAlternativeName

X509Extension

X.509 version 3 extension

X509ExtensionParser

X509Extension parser builder

Enums

DistributionPointName

GeneralName

Represents a GeneralName as defined in RFC5280. There is no support X.400 addresses and EDIPartyName.

ParsedExtension

Functions

parse_ct_signed_certificate_timestamp

Parses as single Signed Certificate Timestamp entry

parse_ct_signed_certificate_timestamp_list

Parses a list of Signed Certificate Timestamp entries

Type Aliases

CertificatePolicies